move entity files to their own folder
