+++ src/main/java/com/stucko09/steam_aggregator/model/entity/GamePlaytimeRecord.java
@@ -1,4 +1,4 @@
-package com.stucko09.steam_aggregator.model;
+package com.stucko09.steam_aggregator.model.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.ManyToOne;
