+++ src/main/java/com/stucko09/steam_aggregator/repository/AppUserRepository.java
@@ -2,7 +2,7 @@ package com.stucko09.steam_aggregator.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.stucko09.steam_aggregator.model.AppUser;
+import com.stucko09.steam_aggregator.model.entity.AppUser;
 
 public interface AppUserRepository extends CrudRepository<AppUser, Long> {
     public AppUser findBySteamUserId(Long steamUserId);
