+++ src/main/java/com/stucko09/steam_aggregator/repository/GamePlaytimeRecordRepository.java
@@ -2,9 +2,9 @@ package com.stucko09.steam_aggregator.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.stucko09.steam_aggregator.model.AppUser;
-import com.stucko09.steam_aggregator.model.GamePlaytimeRecord;
-import com.stucko09.steam_aggregator.model.GameRecord;
+import com.stucko09.steam_aggregator.model.entity.AppUser;
+import com.stucko09.steam_aggregator.model.entity.GamePlaytimeRecord;
+import com.stucko09.steam_aggregator.model.entity.GameRecord;
 
 public interface GamePlaytimeRecordRepository extends CrudRepository<GamePlaytimeRecord, Long> {
     public boolean existsByAppUser(AppUser appUser);
