+++ src/main/java/com/stucko09/steam_aggregator/repository/GameRecordRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.stucko09.steam_aggregator.model.GameRecord;
+import com.stucko09.steam_aggregator.model.entity.GameRecord;
 
 public interface GameRecordRepository extends CrudRepository<GameRecord, Long> {
     public boolean existsBySteamAppId(Long steamAppId);
