+++ src/main/java/com/stucko09/steam_aggregator/service/UserStatsService.java
@@ -3,9 +3,9 @@ package com.stucko09.steam_aggregator.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.stucko09.steam_aggregator.model.AppUser;
-import com.stucko09.steam_aggregator.model.GameRecord;
-import com.stucko09.steam_aggregator.model.UserOwnedGameRecord;
+import com.stucko09.steam_aggregator.model.entity.AppUser;
+import com.stucko09.steam_aggregator.model.entity.GameRecord;
+import com.stucko09.steam_aggregator.model.entity.UserOwnedGameRecord;
 import com.stucko09.steam_aggregator.model.steam.SteamGamePlaytimeRecord;
 import com.stucko09.steam_aggregator.model.steam.SteamGetOwnedGamesResponse;
 import com.stucko09.steam_aggregator.model.steam.SteamGetRecentGamesResponse;
